requirements.txt
torch
torchvision
transformers
diffusers
pillow
numpy
pandas
hpsv2
ImageReward 